ZON-6478: Define cache in zeit.cms for content
