+++ core/src/zeit/content/cp/configure.zcml
@@ -177,7 +177,7 @@
     type="zeit.cms.content.interfaces.IAddableContent"
     />
 
-  <class class=".centerpage.writeabledict">
+  <class class="zeit.cms.content.cache.writeabledict">
     <require
       interface="zope.interface.common.mapping.IFullMapping"
       permission="zope.View" />
